# api/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "教室排课、教室资源管理和冲突检测 RESTful API。",
        "title": "教室排课助手 API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/schedule-drafts": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-drafts"
                ],
                "summary": "List schedule drafts, newest first",
                "parameters": [
                    {
                        "type": "string",
                        "description": "queued|running|succeeded|failed|canceled",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "The draft id is returned immediately and timetable generation runs in the background.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-drafts"
                ],
                "summary": "Submit scheduling conditions and create a draft",
                "parameters": [
                    {
                        "description": "scheduling input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.GenerateScheduleRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedule-drafts/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-drafts"
                ],
                "summary": "Get a draft with progress, current conflicts and failure reason",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "draft id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedule-drafts/{id}/cancel": {
            "post": {
                "description": "Canceling a draft already in a terminal state returns that state unchanged.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-drafts"
                ],
                "summary": "Cancel a queued or running draft (idempotent)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "draft id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
        "/api/v1/schedule-drafts/{id}/result": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "schedule-drafts"
                ],
                "summary": "Get the full timetable, conflicts and summary of a successful draft",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "draft id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_gbschedule_gbschedule_internal_dto.CourseRequirement": {
            "type": "object",
            "required": [
                "course_id",
                "weekly_periods"
            ],
            "properties": {
                "class_id": {
                    "type": "integer"
                },
                "consecutive": {
                    "type": "boolean"
                },
                "course_id": {
                    "type": "integer"
                },
                "teacher_id": {
                    "type": "integer"
                },
                "weekly_periods": {
                    "type": "integer",
                    "maximum": 40,
                    "minimum": 1
                }
            }
        },
...
        "github_com_gbschedule_gbschedule_internal_dto.GenerateScheduleRequest": {
            "type": "object",
            "required": [
                "courses",
                "days_per_week",
                "periods_per_day",
                "weeks"
            ],
            "properties": {
                "class_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "classroom_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "courses": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_gbschedule_gbschedule_internal_dto.CourseRequirement"
                    }
                },
                "days_per_week": {
                    "type": "integer",
                    "maximum": 7,
                    "minimum": 1
                },
                "periods_per_day": {
                    "type": "integer",
                    "maximum": 20,
                    "minimum": 1
                },
                "semester": {
                    "type": "string",
                    "maxLength": 128
                },
                "teacher_ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "weeks": {
                    "type": "integer",
                    "maximum": 30,
                    "minimum": 1
                }
            }
        },
...
        "github_com_gbschedule_gbschedule_internal_dto.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    }
}